+++ app/(auth)/forgetPassword.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import ForgetPasswordScreen from '../../components/ForgetPasswordScreen'
+
+export default function forgetPassword() {
+  return (
+    <ForgetPasswordScreen />
+  )
+}
